requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
dataclasses==0.8
Flask==2.0.3
idna==3.8
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.1
MarkupSafe==2.0.1
numpy==1.19.5
pandas==1.1.5
Pillow==8.4.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.27.1
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
threadpoolctl==3.1.0
torch==1.10.2
torchvision==0.11.3
typing_extensions==4.1.1
urllib3==1.26.20
Werkzeug==2.0.3
wincertstore==0.2
xgboost==1.5.2
zipp==3.6.0
